move autolink to separate file
